harnesses: implement 8 concrete harness classes, retire agents/
